cmake version requirement 3.8
cmake requires higer version to ensure c++17 is available
Removed unused codes and merged messageing / thread related files
