--- modules/deprecated/linkable.h
@@ -1,30 +1,0 @@
-﻿// ---------------------------------------------------------------------------
-//
-//  Author  : github.com/luncliff (luncliff@gmail.com)
-//  License : CC BY 4.0
-//
-//  Note
-//      Macro helper for DLL interface declaration
-//
-// ---------------------------------------------------------------------------
-
-#ifndef LINKABLE_DLL_MACRO
-#define LINKABLE_DLL_MACRO
-
-#ifdef _MSC_VER // MSVC
-#define _HIDDEN_
-#ifdef _WINDLL
-#define _INTERFACE_ __declspec(dllexport)
-#else
-#define _INTERFACE_ __declspec(dllimport)
-#endif
-
-#elif __GNUC__ || __clang__ // GCC or Clang
-#define _INTERFACE_ __attribute__((visibility("default")))
-#define _HIDDEN_ __attribute__((visibility("hidden")))
-
-#else
-#error "unexpected compiler"
-
-#endif // compiler check
-#endif // LINKABLE_DLL_MACRO
