+++ modules/thread/get-code.cmake
@@ -3,13 +3,14 @@
 #   Author  : github.com/luncliff (luncliff@gmail.com)
 #
 # ---------------------------------------------------------------------------
-cmake_minimum_required(VERSION 3.5)
-
-# message(STATUS ${CMAKE_CXX_PLATFORM_ID})
 
 target_sources(${PROJECT_NAME}
 PRIVATE
     thread/types.h
     thread/compatible.cpp
-    thread/${PLATFORM}.cpp
+    thread/registry_${PLATFORM}.cpp
+
+    thread/queue.h
+    thread/queue.cpp
+    thread/operations.cpp
 )
