+++ test/generic/generator.cpp
@@ -2,8 +2,10 @@
 //  Author  : github.com/luncliff (luncliff@gmail.com)
 //  License : CC BY 4.0
 //
-#include "./test.h"
 #include <catch.hpp>
+
+#include <array>
+#include <numeric>
 
 #include <coroutine/enumerable.hpp>
