+++ test/thread/switch_to.cpp
@@ -2,7 +2,6 @@
 //  Author  : github.com/luncliff (luncliff@gmail.com)
 //  License : CC BY 4.0
 //
-#include "./test.h"
 #include <catch.hpp>
 
 #include <thread>
